+++ api/src/fftfpga.c
@@ -9,28 +9,22 @@
 #include <CL/cl_ext_intelfpga.h> // to disable interleaving & transfer data to specific banks - CL_CHANNEL_1_INTELFPGA
 #include "CL/opencl.h"
 
+#include "fpga_state.h"
 #include "fftfpga/fftfpga.h"
 #include "svm.h"
 #include "opencl_utils.h"
 #include "misc.h"
 
-#ifndef KERNEL_VARS
-#define KERNEL_VARS
-static cl_platform_id platform = NULL;
-static cl_device_id *devices;
-static cl_device_id device = NULL;
-static cl_context context = NULL;
-static cl_program program = NULL;
-static cl_command_queue queue1 = NULL, queue2 = NULL, queue3 = NULL;
-static cl_command_queue queue4 = NULL, queue5 = NULL, queue6 = NULL;
-static cl_command_queue queue7 = NULL, queue8 = NULL;
-
+cl_platform_id platform = NULL;
+cl_device_id *devices;
+cl_device_id device = NULL;
+cl_context context = NULL;
+cl_program program = NULL;
+cl_command_queue queue1 = NULL, queue2 = NULL, queue3 = NULL;
+cl_command_queue queue4 = NULL, queue5 = NULL, queue6 = NULL;
+cl_command_queue queue7 = NULL, queue8 = NULL;
 //static int svm_handle;
-static bool svm_enabled = false;
-#endif
-
-static void queue_setup();
-void queue_cleanup();
+bool svm_enabled = false;
 
 /** 
  * @brief Allocate memory of double precision complex floating points
@@ -155,1383 +149,49 @@ void fpga_final(){
 }
 
 /**
- * \brief  compute an out-of-place double precision complex 1D-FFT on the FPGA
- * \param  N    : integer pointer to size of FFT3d  
- * \param  inp  : double2 pointer to input data of size N
- * \param  out  : double2 pointer to output data of size N
- * \param  inv  : int toggle to activate backward FFT
- * \param  batch : number of batched executions of 1D FFT
- * \return fpga_t : time taken in milliseconds for data transfers and execution
+ * \brief Create a command queue for each kernel
  */
-fpga_t fftfpga_c2c_1d(int N, double2 *inp, double2 *out, bool inv, int batch){
-  fpga_t fft_time = {0.0, 0.0, 0.0, 0};
-  cl_kernel fetch_kernel = NULL, fft_kernel = NULL;
+void queue_setup(){
   cl_int status = 0;
-
-  // if N is not a power of 2
-  if(inp == NULL || out == NULL || ((N & (N-1)) !=0)){
-    return fft_time;
-  }
-
-#ifdef VERBOSE
-  printf("Launching%s FFT transform of %d batches \n", inv ? " inverse":"", batch);
-#endif
-
-  queue_setup();
-
-  cl_mem d_inData, d_outData;
-  d_inData = clCreateBuffer(context, CL_MEM_READ_WRITE, sizeof(double2) * N * batch, NULL, &status);
-  checkError(status, "Failed to allocate input device buffer\n");
-
-  d_outData = clCreateBuffer(context, CL_MEM_READ_WRITE | CL_CHANNEL_2_INTELFPGA, sizeof(double2) * N * batch, NULL, &status);
-  checkError(status, "Failed to allocate output device buffer\n");
-
-  // Copy data from host to device
-  fft_time.pcie_write_t = getTimeinMilliSec();
-
-  status = clEnqueueWriteBuffer(queue1, d_inData, CL_TRUE, 0, sizeof(double2) * N * batch, inp, 0, NULL, NULL);
-
-  status = clFinish(queue1);
-  checkError(status, "failed to finish writing buffer using PCIe");
-
-  fft_time.pcie_write_t = getTimeinMilliSec() - fft_time.pcie_write_t;
-  checkError(status, "Failed to copy data to device");
-
-  // Can't pass bool to device, so convert it to int
-  int inverse_int = (int)inv;
-
-  // Create Kernels - names must match the kernel name in the original CL file
-  fetch_kernel = clCreateKernel(program, "fetch", &status);
-  checkError(status, "Failed to create fetch kernel");
-
-  fft_kernel = clCreateKernel(program, "fft1d", &status);
-  checkError(status, "Failed to create fft1d kernel");
-  // Set the kernel arguments
-  // from here
-  status = clSetKernelArg(fetch_kernel, 0, sizeof(cl_mem), (void *)&d_inData);
-  checkError(status, "Failed to set fetch_kernel arg 0");
-  status = clSetKernelArg(fft_kernel, 0, sizeof(cl_mem), (void *)&d_outData);
-  checkError(status, "Failed to set fft_kernel arg 0");
-  status = clSetKernelArg(fft_kernel, 1, sizeof(cl_int), (void*)&batch);
-  checkError(status, "Failed to set fft_kernel arg 1");
-  status = clSetKernelArg(fft_kernel, 2, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set fft_kernel arg 2");
-
-  printf(inverse_int ? "\tInverse FFT" : "\tFFT");
-  printf(" kernel initialization is complete.\n");
-
-  size_t ls = N/8;
-  size_t gs = batch * ls;
-
-  // Measure execution time
-  fft_time.exec_t = getTimeinMilliSec();
-
-  // FFT1d kernel is the SWI kernel
-  status = clEnqueueTask(queue1, fft_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fft1d kernel");
-
-  status = clEnqueueNDRangeKernel(queue2, fetch_kernel, 1, NULL, &gs, &ls, 0, NULL, NULL);
-  checkError(status, "Failed to launch fetch kernel");
-  
-  // Wait for command queue to complete pending events
-  status = clFinish(queue1);
-  checkError(status, "Failed to finish queue1");
-  status = clFinish(queue2);
-  checkError(status, "Failed to finish queue2");
-  
-  // Record execution time
-  fft_time.exec_t = getTimeinMilliSec() - fft_time.exec_t;
-
-  // Copy results from device to host
-  fft_time.pcie_read_t = getTimeinMilliSec();
-  status = clEnqueueReadBuffer(queue1, d_outData, CL_TRUE, 0, sizeof(float2) * N * batch, out, 0, NULL, NULL);
-
-  status = clFinish(queue1);
-  checkError(status, "failed to finish reading buffer using PCIe");
-
-  fft_time.pcie_read_t = getTimeinMilliSec() - fft_time.pcie_read_t;
-  checkError(status, "Failed to copy data from device");
-
-  // Cleanup
-  if (d_inData)
-  	clReleaseMemObject(d_inData);
-  if (d_outData) 
-	  clReleaseMemObject(d_outData);
-  if(fetch_kernel)
-    clReleaseKernel(fetch_kernel);
-  if(fft_kernel)
-    clReleaseKernel(fft_kernel);
-  queue_cleanup();
-  fft_time.valid = 1;
-  return fft_time;
+  // Create one command queue for each kernel.
+  queue1 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
+  checkError(status, "Failed to create command queue1");
+  queue2 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
+  checkError(status, "Failed to create command queue2");
+  queue3 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
+  checkError(status, "Failed to create command queue3");
+  queue4 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
+  checkError(status, "Failed to create command queue4");
+  queue5 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
+  checkError(status, "Failed to create command queue5");
+  queue6 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
+  checkError(status, "Failed to create command queue6");
+  queue7 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
+  checkError(status, "Failed to create command queue6");
+  queue8 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
+  checkError(status, "Failed to create command queue6");
 }
 
 /**
- * \brief  compute an out-of-place single precision complex 1D-FFT on the FPGA
- * \param  N    : integer pointer to size of FFT3d  
- * \param  inp  : float2 pointer to input data of size N
- * \param  out  : float2 pointer to output data of size N
- * \param  inv  : true for backward transforms
- * \param  batch : number of batched executions of 1D FFT
- * \return fpga_t : time taken in milliseconds for data transfers and execution
+ * \brief Release all command queues
  */
-fpga_t fftfpgaf_c2c_1d(int N, float2 *inp, float2 *out, bool inv, int batch){
-
-  fpga_t fft_time = {0.0, 0.0, 0.0, 0};
-  cl_kernel kernel1 = NULL, kernel2 = NULL;
-  cl_int status = 0;
-
-  // if N is not a power of 2
-  if(inp == NULL || out == NULL || ( (N & (N-1)) !=0)){
-    return fft_time;
-  }
-
-#ifdef VERBOSE
-  printf("Launching%s FFT transform for %d batch \n", inv ? " inverse":"", batch);
-#endif
-
-  queue_setup();
-
-  cl_mem d_inData, d_outData;
-  printf("Launching%s FFT transform for %d batch \n", inv ? " inverse":"", batch);
-
-  // Create device buffers - assign the buffers in different banks for more efficient memory access 
-  d_inData = clCreateBuffer(context, CL_MEM_READ_ONLY, sizeof(float2) * N * batch, NULL, &status);
-  checkError(status, "Failed to allocate input device buffer\n");
-
-  d_outData = clCreateBuffer(context, CL_MEM_WRITE_ONLY | CL_CHANNEL_2_INTELFPGA, sizeof(float2) * N * batch, NULL, &status);
-  checkError(status, "Failed to allocate output device buffer\n");
-
-  // Copy data from host to device
-  fft_time.pcie_write_t = getTimeinMilliSec();
-
-  status = clEnqueueWriteBuffer(queue1, d_inData, CL_TRUE, 0, sizeof(float2) * N * batch, inp, 0, NULL, NULL);
-
-  status = clFinish(queue1);
-  checkError(status, "failed to finish writing buffer using PCIe");
-
-  fft_time.pcie_write_t = getTimeinMilliSec() - fft_time.pcie_write_t;
-  checkError(status, "Failed to copy data to device");
-
-  // Can't pass bool to device, so convert it to int
-  int inverse_int = (int)inv;
-
-  // Create Kernels - names must match the kernel name in the original CL file
-  kernel1 = clCreateKernel(program, "fetch", &status);
-  checkError(status, "Failed to create fetch kernel");
-
-  kernel2 = clCreateKernel(program, "fft1d", &status);
-  checkError(status, "Failed to create fft1d kernel");
-  // Set the kernel arguments
-  status = clSetKernelArg(kernel1, 0, sizeof(cl_mem), (void *)&d_inData);
-  checkError(status, "Failed to set kernel1 arg 0");
-  status = clSetKernelArg(kernel2, 0, sizeof(cl_mem), (void *)&d_outData);
-  checkError(status, "Failed to set kernel arg 0");
-  status = clSetKernelArg(kernel2, 1, sizeof(cl_int), (void*)&batch);
-  checkError(status, "Failed to set kernel arg 1");
-  status = clSetKernelArg(kernel2, 2, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set kernel arg 2");
-
-  printf(inverse_int ? "\tInverse FFT" : "\tFFT");
-  printf(" kernel initialization is complete.\n");
-
-  size_t ls = N/8;
-  size_t gs = batch * ls;
-
-  // Measure execution time
-  fft_time.exec_t = getTimeinMilliSec();
-
-  // Launch the kernel - we launch a single work item hence enqueue a task
-  // FFT1d kernel is the SWI kernel
-  status = clEnqueueTask(queue1, kernel2, 0, NULL, NULL);
-  checkError(status, "Failed to launch fft1d kernel");
-
-  status = clEnqueueNDRangeKernel(queue2, kernel1, 1, NULL, &gs, &ls, 0, NULL, NULL);
-  checkError(status, "Failed to launch fetch kernel");
-  
-  // Wait for command queue to complete pending events
-  status = clFinish(queue1);
-  checkError(status, "Failed to finish queue1");
-  status = clFinish(queue2);
-  checkError(status, "Failed to finish queue2");
-  
-  // Record execution time
-  fft_time.exec_t = getTimeinMilliSec() - fft_time.exec_t;
-
-  // Copy results from device to host
-  fft_time.pcie_read_t = getTimeinMilliSec();
-  status = clEnqueueReadBuffer(queue1, d_outData, CL_TRUE, 0, sizeof(float2) * N * batch, out, 0, NULL, NULL);
-
-  status = clFinish(queue1);
-  checkError(status, "failed to finish reading buffer using PCIe");
-
-  fft_time.pcie_read_t = getTimeinMilliSec() - fft_time.pcie_read_t;
-  checkError(status, "Failed to copy data from device");
-
-  // Cleanup
-  if (d_inData)
-  	clReleaseMemObject(d_inData);
-  if (d_outData) 
-	  clReleaseMemObject(d_outData);
-  if(kernel1)
-    clReleaseKernel(kernel1);
-  if(kernel2)
-    clReleaseKernel(kernel2);
-  queue_cleanup();
-
-  fft_time.valid = 1;
-  return fft_time;
-}
-
-/**
- * \brief  compute an out-of-place single precision complex 1D-FFT on the FPGA using Shared Virtual Memory for data transfers between host's main memory and FPGA
- * \param  N    : integer pointer to size of FFT3d  
- * \param  inp  : float2 pointer to input data of size N
- * \param  out  : float2 pointer to output data of size N
- * \param  inv  : int toggle to activate backward FFT
- * \param  batch : number of batched executions of 1D FFT
- * \return fpga_t : time taken in milliseconds for data transfers and execution
- */
-fpga_t fftfpgaf_c2c_1d_svm(int N, float2 *inp, float2 *out, bool inv, int batch){
-  fpga_t fft_time = {0.0, 0.0, 0.0, 0};
-  cl_int status = 0;
-  int num_pts = N * batch;
-  
-  // if N is not a power of 2
-  if(inp == NULL || out == NULL || ( (N & (N-1)) !=0) || !(svm_enabled)){
-    return fft_time;
-  }
-
-#ifdef VERBOSE
-  printf("Launching%s 1D FFT transform in DDR \n", inv ? " inverse":"");
-#endif
-
-  // Can't pass bool to device, so convert it to int
-  int inverse_int = (int)inv;
-
-  // Setup kernels
-  cl_kernel fetch_kernel = clCreateKernel(program, "fetch", &status);
-  checkError(status, "Failed to create fetch1 kernel");
-  cl_kernel fft_kernel = clCreateKernel(program, "fft1d", &status);
-  checkError(status, "Failed to create fft3da kernel");
-
-  // Setup Queues to the kernels
-  queue_setup();
-
-  // allocate SVM buffers
-  float2 *h_inData, *h_outData;
-  h_inData = (float2 *)clSVMAlloc(context, CL_MEM_READ_ONLY, sizeof(float2) * num_pts, 0);
-  h_outData = (float2 *)clSVMAlloc(context, CL_MEM_WRITE_ONLY, sizeof(float2) * num_pts, 0);
-
-  status = clEnqueueSVMMap(queue1, CL_TRUE, CL_MAP_WRITE, (void *)h_inData, sizeof(float2) * num_pts, 0, NULL, NULL);
-  checkError(status, "Failed to map input data");
-
-  // copy data into h_inData
-  for(size_t i = 0; i < num_pts; i++){
-    h_inData[i].x = inp[i].x;
-    h_inData[i].y = inp[i].y;
-  }
-
-  status = clEnqueueSVMUnmap(queue1, (void *)h_inData, 0, NULL, NULL);
-  checkError(status, "Failed to unmap input data");
-
-  status = clEnqueueSVMMap(queue1, CL_TRUE, CL_MAP_WRITE, (void *)h_outData, sizeof(float2) * num_pts, 0, NULL, NULL);
-  checkError(status, "Failed to map input data");
-
-  // copy data into h_inData
-  for(size_t i = 0; i < num_pts; i++){
-    h_outData[i].x = 0.0;
-    h_outData[i].y = 0.0;
-  }
-
-  status = clEnqueueSVMUnmap(queue1, (void *)h_outData, 0, NULL, NULL);
-  checkError(status, "Failed to unmap input data");
-
-  // write to fetch kernel using SVM based PCIe
-  status = clSetKernelArgSVMPointer(fetch_kernel, 0, (void *)h_inData);
-  checkError(status, "Failed to set fetch kernel arg");
-
-  // kernel transforms and stores to DDR memory
-  status = clSetKernelArgSVMPointer(fft_kernel, 0, (void *)h_outData);
-  checkError(status, "Failed to set store2 kernel arg");
-
-  status=clSetKernelArg(fft_kernel, 1, sizeof(cl_int), (void*)&batch);
-  checkError(status, "Failed to set fft kernel arg");
-
-  status=clSetKernelArg(fft_kernel, 2, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set fft kernel arg");
-
-  size_t ls = N/8;
-  size_t gs = batch * ls;
-
-  fft_time.exec_t = getTimeinMilliSec();
-  status = clEnqueueTask(queue1, fft_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fetch kernel");
-
-  status = clEnqueueNDRangeKernel(queue2, fetch_kernel, 1, NULL, &gs, &ls, 0, NULL, NULL);
-  checkError(status, "Failed to launch fetch kernel");
-
-  status = clFinish(queue2);
-  checkError(status, "failed to finish");
-  status = clFinish(queue1);
-  checkError(status, "failed to finish");
-
-  fft_time.exec_t = getTimeinMilliSec() - fft_time.exec_t;
-
-  status = clEnqueueSVMMap(queue1, CL_TRUE, CL_MAP_READ,
-    (void *)h_outData, sizeof(float2) * num_pts, 0, NULL, NULL);
-  checkError(status, "Failed to map out data");
-
-  for(size_t i = 0; i < num_pts; i++){
-    out[i].x = h_outData[i].x;
-    out[i].y = h_outData[i].y;
-  }
-
-  status = clEnqueueSVMUnmap(queue1, (void *)h_outData, 0, NULL, NULL);
-  checkError(status, "Failed to unmap out data");
-
-  if (h_inData)
-    clSVMFree(context, h_inData);
-  if (h_outData)
-    clSVMFree(context, h_outData);
-
-  queue_cleanup();
-
-  if(fetch_kernel) 
-    clReleaseKernel(fetch_kernel);  
-
-  if(fft_kernel) 
-    clReleaseKernel(fft_kernel);  
-
-  fft_time.valid = 1;
-  return fft_time;
-}
-
-/**
- * \brief  compute an out-of-place single precision complex 2D-FFT using the DDR of the FPGA
- * \param  N    : integer pointer to size of FFT2d  
- * \param  inp  : float2 pointer to input data of size [N * N]
- * \param  out  : float2 pointer to output data of size [N * N]
- * \param  inv  : int toggle to activate backward FFT
- * \param  iter : int toggle to activate backward FFT
- * \return fpga_t : time taken in milliseconds for data transfers and execution
- */
-fpga_t fftfpgaf_c2c_2d_ddr(int N, float2 *inp, float2 *out, bool inv){
-  fpga_t fft_time = {0.0, 0.0, 0.0, 0};
-  cl_kernel fetch_kernel = NULL, fft_kernel = NULL, transpose_kernel = NULL;
-  cl_int status = 0;
-  int mangle_int = 0;
-
-  // if N is not a power of 2
-  if(inp == NULL || out == NULL || ( (N & (N-1)) !=0)){
-    return fft_time;
-  }
-
-#ifdef VERBOSE
-  printf("Launching%s 2d FFT transform \n", inv ? " inverse":"");
-#endif
-
-  queue_setup();
-
-  cl_mem d_inData, d_outData, d_tmp;
-
-  d_inData = clCreateBuffer(context, CL_MEM_READ_ONLY, sizeof(float2) * N * N, NULL, &status);
-  checkError(status, "Failed to allocate input device buffer\n");
-  d_outData = clCreateBuffer(context, CL_MEM_WRITE_ONLY, sizeof(float2) * N * N, NULL, &status);
-  checkError(status, "Failed to allocate output device buffer\n");
-  d_tmp = clCreateBuffer(context, CL_MEM_READ_WRITE | CL_CHANNEL_2_INTELFPGA, sizeof(float2) * N * N, NULL, &status);
-  checkError(status, "Failed to allocate output device buffer\n");
-
-  // Copy data from host to device
-  fft_time.pcie_write_t = getTimeinMilliSec();
-
-  status = clEnqueueWriteBuffer(queue1, d_inData, CL_TRUE, 0, sizeof(float2) * N * N, inp, 0, NULL, NULL);
-
-  status = clFinish(queue1);
-  checkError(status, "failed to finish writing buffer using PCIe");
-
-  fft_time.pcie_write_t = getTimeinMilliSec() - fft_time.pcie_write_t;
-  checkError(status, "Failed to copy data to device");
-
-  // Can't pass bool to device, so convert it to int
-  int inverse_int = (int)inv;
-
-  // Create Kernels - names must match the kernel name in the original CL file
-  fft_kernel = clCreateKernel(program, "fft2d", &status);
-  checkError(status, "Failed to create kernel");
-  fetch_kernel = clCreateKernel(program, "fetch", &status);
-  checkError(status, "Failed to create kernel");
-  transpose_kernel = clCreateKernel(program, "transpose", &status);
-  checkError(status, "Failed to create kernel");
-
-  // Record execution time
-  fft_time.exec_t = getTimeinMilliSec();
-
-  // Loop twice over the kernels
-  for (size_t i = 0; i < 2; i++) {
-
-    // Set the kernel arguments
-    status = clSetKernelArg(fetch_kernel, 0, sizeof(cl_mem), i == 0 ? (void *)&d_inData : (void *)&d_tmp);
-    checkError(status, "Failed to set kernel arg 0");
-    status = clSetKernelArg(fetch_kernel, 1, sizeof(cl_int), (void*)&mangle_int);
-    checkError(status, "Failed to set kernel arg 1");
-    size_t lws_fetch[] = {N};
-    size_t gws_fetch[] = {N * N / 8};
-    status = clEnqueueNDRangeKernel(queue1, fetch_kernel, 1, 0, gws_fetch, lws_fetch, 0, NULL, NULL);
-    checkError(status, "Failed to launch kernel");
-
-    // Launch the fft kernel - we launch a single work item hence enqueue a task
-    status = clSetKernelArg(fft_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-    checkError(status, "Failed to set kernel arg 0");
-    status = clEnqueueTask(queue2, fft_kernel, 0, NULL, NULL);
-    checkError(status, "Failed to launch kernel");
-
-    // Set the kernel arguments
-    status = clSetKernelArg(transpose_kernel, 0, sizeof(cl_mem), i == 0 ? (void *)&d_tmp : (void *)&d_outData);
-    checkError(status, "Failed to set kernel arg 0");
-
-    status = clSetKernelArg(transpose_kernel, 1, sizeof(cl_int), (void*)&mangle_int);
-    checkError(status, "Failed to set kernel arg 1");
-
-    size_t lws_transpose[] = {N};
-    size_t gws_transpose[] = {N * N / 8};
-    status = clEnqueueNDRangeKernel(queue3, transpose_kernel, 1, 0, gws_transpose, lws_transpose, 0, NULL, NULL);
-    checkError(status, "Failed to launch kernel");
-
-    // Wait for all command queues to complete pending events
-    status = clFinish(queue1);
-    checkError(status, "failed to finish");
-    status = clFinish(queue2);
-    checkError(status, "failed to finish");
-    status = clFinish(queue3);
-    checkError(status, "failed to finish");
-  }
-
-  fft_time.exec_t = getTimeinMilliSec() - fft_time.exec_t;
-
-  // Copy results from device to host
-  fft_time.pcie_read_t = getTimeinMilliSec();
-  status = clEnqueueReadBuffer(queue1, d_outData, CL_TRUE, 0, sizeof(float2) * N * N, out, 0, NULL, NULL);
-
-  status = clFinish(queue1);
-  checkError(status, "failed to finish reading buffer using PCIe");
-
-  fft_time.pcie_read_t = getTimeinMilliSec() - fft_time.pcie_read_t;
-  checkError(status, "Failed to copy data from device");
-
-  // Cleanup
-  if (d_inData)
-  	clReleaseMemObject(d_inData);
-  if (d_outData) 
-	  clReleaseMemObject(d_outData);
-  if (d_tmp)
-	  clReleaseMemObject(d_tmp);
-  if(fft_kernel)
-    clReleaseKernel(fft_kernel);
-  if(fetch_kernel)
-    clReleaseKernel(fetch_kernel);
-  if(transpose_kernel)
-    clReleaseKernel(transpose_kernel);
-  queue_cleanup();
-
-  fft_time.valid = 1;
-  return fft_time;
-}
-
-/**
- * \brief  compute an out-of-place single precision complex 2D-FFT using the BRAM of the FPGA
- * \param  N    : integer pointer to size of FFT2d  
- * \param  inp  : float2 pointer to input data of size [N * N]
- * \param  out  : float2 pointer to output data of size [N * N]
- * \param  inv  : int toggle to activate backward FFT
- * \param  interleaving : enable interleaved global memory buffers
- * \return fpga_t : time taken in milliseconds for data transfers and execution
- */
-fpga_t fftfpgaf_c2c_2d_bram(int N, float2 *inp, float2 *out, bool inv, bool interleaving, int how_many){
-  fpga_t fft_time = {0.0, 0.0, 0.0, 0};
-  cl_kernel ffta_kernel = NULL, fftb_kernel = NULL;
-  cl_kernel fetch_kernel = NULL, store_kernel = NULL;
-  cl_kernel transpose_kernel = NULL;
-
-  cl_int status = 0;
-  int num_pts = how_many * N * N;
-
-  // if N is not a power of 2
-  if(inp == NULL || out == NULL || ( (N & (N-1)) !=0)){
-    return fft_time;
-  }
-
-#ifdef VERBOSE
-  printf("Launching%s 3d FFT transform in DDR \n", inv ? " inverse":"");
-#endif
-
-  queue_setup();
-
-  cl_mem_flags flagbuf1, flagbuf2;
-  if(interleaving == 1){
-    flagbuf1 = CL_MEM_READ_WRITE;
-    flagbuf2 = CL_MEM_READ_WRITE;
-  }
-  else{
-    flagbuf1 = CL_MEM_READ_ONLY | CL_CHANNEL_1_INTELFPGA;
-    flagbuf2 = CL_MEM_WRITE_ONLY | CL_CHANNEL_2_INTELFPGA;
-  }
-  
-  // Device memory buffers
-  cl_mem d_inData, d_outData;
-  d_inData = clCreateBuffer(context, flagbuf1, sizeof(float2) * num_pts, NULL, &status);
-  checkError(status, "Failed to allocate input device buffer\n");
-
-  d_outData = clCreateBuffer(context, flagbuf2, sizeof(float2) * num_pts, NULL, &status);
-  checkError(status, "Failed to allocate output device buffer\n");
-
- // Copy data from host to device
-  fft_time.pcie_write_t = getTimeinMilliSec();
-
-  status = clEnqueueWriteBuffer(queue1, d_inData, CL_TRUE, 0, sizeof(float2) * num_pts, inp, 0, NULL, NULL);
-
-  status = clFinish(queue1);
-  checkError(status, "failed to finish");
-
-  fft_time.pcie_write_t = getTimeinMilliSec() - fft_time.pcie_write_t;
-  checkError(status, "Failed to copy data to device");
-
-  // Can't pass bool to device, so convert it to int
-  int inverse_int = (int)inv;
-
-  ffta_kernel = clCreateKernel(program, "fft2da", &status);
-  checkError(status, "Failed to create fft2da kernel");
-
-  fftb_kernel = clCreateKernel(program, "fft2db", &status);
-  checkError(status, "Failed to create fft2db kernel");
-
-  fetch_kernel = clCreateKernel(program, "fetchBitrev", &status);
-  checkError(status, "Failed to create fetch kernel");
-
-  transpose_kernel = clCreateKernel(program, "transpose", &status);
-  checkError(status, "Failed to create transpose1 kernel");
-
-  store_kernel = clCreateKernel(program, "transposeStore", &status);
-  checkError(status, "Failed to create store kernel");
-
-  status = clSetKernelArg(fetch_kernel, 0, sizeof(cl_mem), (void *)&d_inData);
-  checkError(status, "Failed to set fetch kernel arg 0");
-
-  status = clSetKernelArg(fetch_kernel, 1, sizeof(cl_int), (void *)&how_many);
-  checkError(status, "Failed to set fetch kernel arg 1");
-
-  status = clSetKernelArg(ffta_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set ffta kernel arg 0");
-
-  status = clSetKernelArg(ffta_kernel, 1, sizeof(cl_int), (void*)&how_many);
-  checkError(status, "Failed to set ffta kernel arg 1");
-
-  status = clSetKernelArg(transpose_kernel, 0, sizeof(cl_int), (void*)&how_many);
-  checkError(status, "Failed to set transpose kernel arg 0");
-
-  status = clSetKernelArg(fftb_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set fftb kernel arg 0");
-
-  status = clSetKernelArg(fftb_kernel, 1, sizeof(cl_int), (void*)&how_many);
-  checkError(status, "Failed to set fftb kernel arg 1");
-
-  status = clSetKernelArg(store_kernel, 0, sizeof(cl_mem), (void *)&d_outData);
-  checkError(status, "Failed to set store kernel arg");
-
-  status = clSetKernelArg(store_kernel, 1, sizeof(cl_int), (void *)&how_many);
-  checkError(status, "Failed to set store kernel arg");
-
-  fft_time.exec_t = getTimeinMilliSec();
-  status = clEnqueueTask(queue1, fetch_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fetch kernel");
-
-  status = clEnqueueTask(queue2, ffta_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fft kernel");
-
-  status = clEnqueueTask(queue3, transpose_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch transpose1 kernel");
-
-  status = clEnqueueTask(queue4, fftb_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch second fft kernel");
-
-  status = clEnqueueTask(queue5, store_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch store kernel");
-
-  // Wait for all command queues to complete pending events
-  status = clFinish(queue1);
-  checkError(status, "failed to finish queue1");
-  status = clFinish(queue2);
-  checkError(status, "failed to finish queue2");
-  status = clFinish(queue3);
-  checkError(status, "failed to finish queue3");
-  status = clFinish(queue4);
-  checkError(status, "failed to finish queue4");
-  status = clFinish(queue5);
-  checkError(status, "failed to finish queue5");
-  fft_time.exec_t = getTimeinMilliSec() - fft_time.exec_t;
-
-  // Copy results from device to host
-  fft_time.pcie_read_t = getTimeinMilliSec();
-  status = clEnqueueReadBuffer(queue1, d_outData, CL_TRUE, 0, sizeof(float2) * num_pts, out, 0, NULL, NULL);
-
-  status = clFinish(queue1);
-  checkError(status, "failed to finish reading buffer using PCIe");
-
-  fft_time.pcie_read_t = getTimeinMilliSec() - fft_time.pcie_read_t;
-  checkError(status, "Failed to copy data from device");
-
-  queue_cleanup();
-
-  if (d_inData)
-  	clReleaseMemObject(d_inData);
-  if (d_outData) 
-	  clReleaseMemObject(d_outData);
-
-  if(fetch_kernel) 
-    clReleaseKernel(fetch_kernel);  
-
-  if(ffta_kernel) 
-    clReleaseKernel(ffta_kernel);  
-  if(fftb_kernel) 
-    clReleaseKernel(fftb_kernel);  
-
-  if(transpose_kernel) 
-    clReleaseKernel(transpose_kernel);  
-
-  if(store_kernel) 
-    clReleaseKernel(store_kernel);  
-
-  fft_time.valid = 1;
-  return fft_time;
-}
-
-/**
- * \brief  compute an out-of-place single precision complex 2DFFT using the BRAM of the FPGA and Shared Virtual Memory for Host to Device Communication
- * \param  N    : integer pointer to size of FFT2d  
- * \param  inp  : float2 pointer to input data of size [N * N]
- * \param  out  : float2 pointer to output data of size [N * N]
- * \param  inv  : int toggle to activate backward FFT
- * \return fpga_t : time taken in milliseconds for data transfers and execution
- */
-fpga_t fftfpgaf_c2c_2d_bram_svm(int N, float2 *inp, float2 *out, bool inv, int how_many){
-  fpga_t fft_time = {0.0, 0.0, 0.0, 0};
-  cl_int status = 0;
-  int num_pts = how_many * N * N;
-
-  cl_kernel ffta_kernel = NULL, fftb_kernel = NULL;
-  cl_kernel fetch_kernel = NULL, store_kernel = NULL;
-  cl_kernel transpose_kernel = NULL;
-
-  // if N is not a power of 2
-  if(inp == NULL || out == NULL || ( (N & (N-1)) !=0) || (!svm_enabled)){
-    return fft_time;
-  }
-
-#ifdef VERBOSE
-  printf("Launching%s 2d FFT transform in BRAM using SVM\n", inv ? " inverse":"");
-#endif
-
-  queue_setup();
-
-  // allocate SVM buffers
-  float2 *h_inData, *h_outData;
-  h_inData = (float2 *)clSVMAlloc(context, CL_MEM_READ_ONLY, sizeof(float2) * num_pts, 0);
-  h_outData = (float2 *)clSVMAlloc(context, CL_MEM_WRITE_ONLY, sizeof(float2) * num_pts, 0);
-
-  status = clEnqueueSVMMap(queue1, CL_TRUE, CL_MAP_WRITE, (void *)h_inData, sizeof(float2) * num_pts, 0, NULL, NULL);
-  checkError(status, "Failed to map input data");
-
-  // copy data into h_inData
-  for(size_t i = 0; i < num_pts; i++){
-    h_inData[i].x = inp[i].x;
-    h_inData[i].y = inp[i].y;
-  }
-
-  status = clEnqueueSVMUnmap(queue1, (void *)h_inData, 0, NULL, NULL);
-  checkError(status, "Failed to unmap input data");
-
-  status = clEnqueueSVMMap(queue1, CL_TRUE, CL_MAP_WRITE, (void *)h_outData, sizeof(float2) * num_pts, 0, NULL, NULL);
-  checkError(status, "Failed to map input data");
-
-  // copy data into h_inData
-  for(size_t i = 0; i < num_pts; i++){
-    h_outData[i].x = 0.0;
-    h_outData[i].y = 0.0;
-  }
-
-  status = clEnqueueSVMUnmap(queue1, (void *)h_outData, 0, NULL, NULL);
-  checkError(status, "Failed to unmap input data");
-
-  // Can't pass bool to device, so convert it to int
-  int inverse_int = (int)inv;
-
-  ffta_kernel = clCreateKernel(program, "fft2da", &status);
-  checkError(status, "Failed to create fft2da kernel");
-
-  fftb_kernel = clCreateKernel(program, "fft2db", &status);
-  checkError(status, "Failed to create fft2db kernel");
-
-  fetch_kernel = clCreateKernel(program, "fetchBitrev", &status);
-  checkError(status, "Failed to create fetch kernel");
-
-  transpose_kernel = clCreateKernel(program, "transpose", &status);
-  checkError(status, "Failed to create transpose1 kernel");
-
-  store_kernel = clCreateKernel(program, "transposeStore", &status);
-  checkError(status, "Failed to create store kernel");
-
-  // write to fetch kernel using SVM based PCIe
-  status = clSetKernelArgSVMPointer(fetch_kernel, 0, (void *)h_inData);
-  checkError(status, "Failed to set fetch1 kernel arg");
-
-  status = clSetKernelArg(fetch_kernel, 1, sizeof(cl_int), (void *)&how_many);
-  checkError(status, "Failed to set fetch kernel arg 1");
-
-  status = clSetKernelArg(ffta_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set ffta kernel arg 0");
-
-  status = clSetKernelArg(ffta_kernel, 1, sizeof(cl_int), (void*)&how_many);
-  checkError(status, "Failed to set ffta kernel arg 1");
-
-  status = clSetKernelArg(transpose_kernel, 0, sizeof(cl_int), (void*)&how_many);
-  checkError(status, "Failed to set transpose kernel arg 0");
-
-  status = clSetKernelArg(fftb_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set fftb kernel arg 0");
-
-  status = clSetKernelArg(fftb_kernel, 1, sizeof(cl_int), (void*)&how_many);
-  checkError(status, "Failed to set fftb kernel arg 1");
-
-  // kernel stores using SVM based PCIe to host
-  status = clSetKernelArgSVMPointer(store_kernel, 0, (void*)h_outData);
-  checkError(status, "Failed to set store2 kernel arg");
-
-  status = clSetKernelArg(store_kernel, 1, sizeof(cl_int), (void *)&how_many);
-  checkError(status, "Failed to set store kernel arg");
-
-  fft_time.exec_t = getTimeinMilliSec();
-  status = clEnqueueTask(queue1, fetch_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fetch kernel");
-
-  status = clEnqueueTask(queue2, ffta_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fft kernel");
-
-  status = clEnqueueTask(queue3, transpose_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch transpose1 kernel");
-
-  status = clEnqueueTask(queue4, fftb_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch second fft kernel");
-
-  status = clEnqueueTask(queue5, store_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch store kernel");
-
-  // Wait for all command queues to complete pending events
-  status = clFinish(queue5);
-  checkError(status, "failed to finish queue5");
-  status = clFinish(queue1);
-  checkError(status, "failed to finish queue1");
-  status = clFinish(queue2);
-  checkError(status, "failed to finish queue2");
-  status = clFinish(queue3);
-  checkError(status, "failed to finish queue3");
-  status = clFinish(queue4);
-  checkError(status, "failed to finish queue4");
-  fft_time.exec_t = getTimeinMilliSec() - fft_time.exec_t;
-
- status = clEnqueueSVMMap(queue1, CL_TRUE, CL_MAP_READ,
-    (void *)h_outData, sizeof(float2) * num_pts, 0, NULL, NULL);
-  checkError(status, "Failed to map out data");
-
-  for(size_t i = 0; i < num_pts; i++){
-    out[i].x = h_outData[i].x;
-    out[i].y = h_outData[i].y;
-  }
-
-  status = clEnqueueSVMUnmap(queue1, (void *)h_outData, 0, NULL, NULL);
-  checkError(status, "Failed to unmap out data");
-
-  if (h_inData)
-    clSVMFree(context, h_inData);
-  if (h_outData)
-    clSVMFree(context, h_outData);
-
-  queue_cleanup();
-
-  if(fetch_kernel) 
-    clReleaseKernel(fetch_kernel);  
-  if(ffta_kernel) 
-    clReleaseKernel(ffta_kernel);  
-  if(fftb_kernel) 
-    clReleaseKernel(fftb_kernel);  
-  if(transpose_kernel) 
-    clReleaseKernel(transpose_kernel);  
-  if(store_kernel) 
-    clReleaseKernel(store_kernel);  
-
-  fft_time.valid = 1;
-  return fft_time;
-}
-
-/**
- * \brief  compute an out-of-place single precision complex 3D-FFT using the BRAM of the FPGA
- * \param  N    : integer pointer addressing the size of FFT3d  
- * \param  inp  : float2 pointer to input data of size [N * N * N]
- * \param  out  : float2 pointer to output data of size [N * N * N]
- * \param  inv  : int toggle to activate backward FFT
- * \param  interleaving : 1 if using burst interleaved global memory buffers
- * \return fpga_t : time taken in milliseconds for data transfers and execution
- */
-fpga_t fftfpgaf_c2c_3d_bram(int N, float2 *inp, float2 *out, bool inv, bool interleaving) {
-  fpga_t fft_time = {0.0, 0.0, 0.0, 0};
-  cl_int status = 0;
-
-  cl_kernel fft3da_kernel = NULL, fft3db_kernel = NULL, fft3dc_kernel = NULL;
-  cl_kernel fetch_kernel = NULL, store_kernel = NULL;
-  cl_kernel transpose_kernel = NULL, transpose3d_kernel = NULL;
-
-  // if N is not a power of 2
-  if(inp == NULL || out == NULL || ( (N & (N-1)) !=0)){
-    return fft_time;
-  }
-
-#ifdef VERBOSE
-  printf("Launching%s 3d FFT transform \n", inv ? " inverse":"");
-#endif
-
-  queue_setup();
-
-  cl_mem_flags flagbuf1, flagbuf2;
-  if(interleaving){
-    flagbuf1 = CL_MEM_READ_WRITE;
-    flagbuf2 = CL_MEM_READ_WRITE;
-  }
-  else{
-    flagbuf1 = CL_MEM_READ_ONLY | CL_CHANNEL_1_INTELFPGA;
-    flagbuf2 = CL_MEM_WRITE_ONLY | CL_CHANNEL_2_INTELFPGA;
-  }
-
-  // Device memory buffers
-  cl_mem d_inData, d_outData;
-  d_inData = clCreateBuffer(context, flagbuf1, sizeof(float2) * N * N * N, NULL, &status);
-  checkError(status, "Failed to allocate input device buffer\n");
-  d_outData = clCreateBuffer(context, flagbuf2, sizeof(float2) * N * N * N, NULL, &status);
-  checkError(status, "Failed to allocate output device buffer\n");
-
- // Copy data from host to device
-  fft_time.pcie_write_t = getTimeinMilliSec();
-
-  status = clEnqueueWriteBuffer(queue1, d_inData, CL_TRUE, 0, sizeof(float2) * N * N * N, inp, 0, NULL, NULL);
-
-  status = clFinish(queue1);
-  checkError(status, "failed to finish");
-
-  fft_time.pcie_write_t = getTimeinMilliSec() - fft_time.pcie_write_t;
-  checkError(status, "Failed to copy data to device");
-
-  // Can't pass bool to device, so convert it to int
-  int inverse_int = (int)inv;
-
-  // Create the kernel - name passed in here must match kernel name in the
-  // original CL file, that was compiled into an AOCX file using the AOC tool
-  fetch_kernel = clCreateKernel(program, "fetch", &status);
-  checkError(status, "Failed to create fetch kernel");
-
-  fft3da_kernel = clCreateKernel(program, "fft3da", &status);
-  checkError(status, "Failed to create fft3da kernel");
-
-  transpose_kernel = clCreateKernel(program, "transpose2d", &status);
-  checkError(status, "Failed to create transpose kernel");
-
-  fft3db_kernel = clCreateKernel(program, "fft3db", &status);
-  checkError(status, "Failed to create fft3db kernel");
-
-  transpose3d_kernel = clCreateKernel(program, "transpose3D", &status);
-  checkError(status, "Failed to create transpose3D kernel");
-
-  fft3dc_kernel = clCreateKernel(program, "fft3dc", &status);
-  checkError(status, "Failed to create fft3dc kernel");
-
-  store_kernel = clCreateKernel(program, "store", &status);
-  checkError(status, "Failed to create store kernel");
-
-  status = clSetKernelArg(fetch_kernel, 0, sizeof(cl_mem), (void *)&d_inData);
-  checkError(status, "Failed to set fetch kernel arg 0");
-  status = clSetKernelArg(fft3da_kernel, 0, sizeof(cl_int),(void*)&inverse_int);
-  checkError(status, "Failed to set fft3da kernel arg 0");
-  status = clSetKernelArg(fft3db_kernel, 0, sizeof(cl_int),(void*)&inverse_int);
-  checkError(status, "Failed to set fft3db_kernel arg 0");
-  status = clSetKernelArg(fft3dc_kernel, 0, sizeof(cl_int),(void*)&inverse_int);
-  checkError(status, "Failed to set fft3dc_kernel arg 0");
-  status = clSetKernelArg(store_kernel, 0, sizeof(cl_mem), (void *)&d_outData);
-  checkError(status, "Failed to set store kernel arg 0");
-
-  fft_time.exec_t = getTimeinMilliSec();
-  status = clEnqueueTask(queue1, fetch_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fetch kernel");
-
-  status = clEnqueueTask(queue2, fft3da_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fft kernel");
-
-  status = clEnqueueTask(queue3, transpose_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch transpose kernel");
-
-  status = clEnqueueTask(queue4, fft3db_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch second fft kernel");
-
-  status = clEnqueueTask(queue5, transpose3d_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch second transpose kernel");
-
-  status = clEnqueueTask(queue6, fft3dc_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch third fft kernel");
-
-  status = clEnqueueTask(queue7, store_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch store transpose kernel");
-
-  // Wait for all command queues to complete pending events
-  status = clFinish(queue1);
-  checkError(status, "failed to finish");
-  status = clFinish(queue2);
-  checkError(status, "failed to finish");
-  status = clFinish(queue3);
-  checkError(status, "failed to finish");
-  status = clFinish(queue4);
-  checkError(status, "failed to finish");
-  status = clFinish(queue5);
-  checkError(status, "failed to finish");
-  status = clFinish(queue6);
-  checkError(status, "failed to finish");
-  status = clFinish(queue7);
-  checkError(status, "failed to finish");
-
-  fft_time.exec_t = getTimeinMilliSec() - fft_time.exec_t;
-
-  // Copy results from device to host
-  fft_time.pcie_read_t = getTimeinMilliSec();
-  status = clEnqueueReadBuffer(queue1, d_outData, CL_TRUE, 0, sizeof(float2) * N * N * N, out, 0, NULL, NULL);
-
-  status = clFinish(queue1);
-  checkError(status, "failed to finish reading buffer using PCIe");
-
-  fft_time.pcie_read_t = getTimeinMilliSec() - fft_time.pcie_read_t;
-  checkError(status, "Failed to copy data from device");
-
-  queue_cleanup();
-
-  if (d_inData)
-  	clReleaseMemObject(d_inData);
-  if (d_outData) 
-	  clReleaseMemObject(d_outData);
-
-  if(fetch_kernel) 
-    clReleaseKernel(fetch_kernel);  
-  if(fft3da_kernel) 
-    clReleaseKernel(fft3da_kernel);  
-  if(fft3db_kernel) 
-    clReleaseKernel(fft3db_kernel);  
-  if(fft3dc_kernel) 
-    clReleaseKernel(fft3dc_kernel);  
-  if(transpose_kernel) 
-    clReleaseKernel(transpose_kernel);  
-  if(transpose3d_kernel) 
-    clReleaseKernel(transpose3d_kernel); 
-  if(store_kernel) 
-    clReleaseKernel(store_kernel);  
-
-  fft_time.valid = 1;
-  return fft_time;
-}
-
-/**
- * \brief  compute an out-of-place single precision complex 3D FFT using the DDR for 3D Transpose where the data access between the host and the FPGA is using Shared Virtual Memory (SVM)
- * \param  N    : integer pointer addressing the size of FFT3d  
- * \param  inp  : float2 pointer to input data of size [N * N * N]
- * \param  out  : float2 pointer to output data of size [N * N * N]
- * \param  inv  : int toggle to activate backward FFT
- * \param  interleaving : 1 if using burst interleaved global memory buffers
- * \return fpga_t : time taken in milliseconds for data transfers and execution
- */
-fpga_t fftfpgaf_c2c_3d_ddr_svm(int N, float2 *inp, float2 *out, bool inv) {
-  fpga_t fft_time = {0.0, 0.0, 0.0, 0};
-  cl_int status = 0;
-  int num_pts = N * N * N;
-  
-  /*
-  const char* board_name;
-  int *bytes;
-  aocl_mmd_offline_info_t info_id;
-  info_id = AOCL_MMD_BOARD_NAMES;
-  aocl_mmd_get_offline_info(info_id, sizeof(char*), &board_name, size_t(int));
-
-  svm_handle = aocl_mmd_open(board_name);
-  if(svm_handle < 0 ){
-    return NULL;
-  }
-  return aocl_mmd_shared_mem_alloc(svm_handle, sz, inData, device_ptr);
-  */
-
-  // if N is not a power of 2
-  if(inp == NULL || out == NULL || ( (N & (N-1)) !=0) || !(svm_enabled)){
-    return fft_time;
-  }
-
-#ifdef VERBOSE
-  printf("Launching%s 3d FFT transform in DDR \n", inv ? " inverse":"");
-#endif
-
-  // Can't pass bool to device, so convert it to int
-  int inverse_int = (int)inv;
-
-  // Setup kernels
-  cl_kernel fetch1_kernel = clCreateKernel(program, "fetchBitrev1", &status);
-  checkError(status, "Failed to create fetch1 kernel");
-  cl_kernel ffta_kernel = clCreateKernel(program, "fft3da", &status);
-  checkError(status, "Failed to create fft3da kernel");
-  cl_kernel transpose_kernel = clCreateKernel(program, "transpose", &status);
-  checkError(status, "Failed to create transpose kernel");
-  cl_kernel fftb_kernel = clCreateKernel(program, "fft3db", &status);
-  checkError(status, "Failed to create fft3db kernel");
-  cl_kernel store1_kernel = clCreateKernel(program, "transposeStore1", &status);
-  checkError(status, "Failed to create store1 kernel");
-
-  cl_kernel fetch2_kernel = clCreateKernel(program, "fetchBitrev2", &status);
-  checkError(status, "Failed to create fetch2 kernel");
-  cl_kernel fftc_kernel = clCreateKernel(program, "fft3dc", &status);
-  checkError(status, "Failed to create fft3dc kernel");
-  cl_kernel store2_kernel = clCreateKernel(program, "transposeStore2", &status);
-  checkError(status, "Failed to create store2 kernel");
-
-  // Setup Queues to the kernels
-  queue_setup();
-
-  // Device memory buffers
-  cl_mem d_inOutData;
-  d_inOutData = clCreateBuffer(context, CL_MEM_READ_WRITE | CL_CHANNEL_1_INTELFPGA, sizeof(float2) * num_pts, NULL, &status);
-  checkError(status, "Failed to allocate output device buffer\n");
-
-  // allocate SVM buffers
-  float2 *h_inData, *h_outData;
-  h_inData = (float2 *)clSVMAlloc(context, CL_MEM_READ_ONLY, sizeof(float2) * num_pts, 0);
-  h_outData = (float2 *)clSVMAlloc(context, CL_MEM_WRITE_ONLY, sizeof(float2) * num_pts, 0);
-
-  status = clEnqueueSVMMap(queue1, CL_TRUE, CL_MAP_WRITE, (void *)h_inData, sizeof(float2) * num_pts, 0, NULL, NULL);
-  checkError(status, "Failed to map input data");
-
-  // copy data into h_inData
-  for(size_t i = 0; i < num_pts; i++){
-    h_inData[i].x = inp[i].x;
-    h_inData[i].y = inp[i].y;
-  }
-
-  status = clEnqueueSVMUnmap(queue1, (void *)h_inData, 0, NULL, NULL);
-  checkError(status, "Failed to unmap input data");
-
-  status = clEnqueueSVMMap(queue1, CL_TRUE, CL_MAP_WRITE, (void *)h_outData, sizeof(float2) * num_pts, 0, NULL, NULL);
-  checkError(status, "Failed to map input data");
-
-  // copy data into h_inData
-  for(size_t i = 0; i < num_pts; i++){
-    h_outData[i].x = 0.0;
-    h_outData[i].y = 0.0;
-  }
-
-  status = clEnqueueSVMUnmap(queue1, (void *)h_outData, 0, NULL, NULL);
-  checkError(status, "Failed to unmap input data");
-
-  // write to fetch kernel using SVM based PCIe
-  status = clSetKernelArgSVMPointer(fetch1_kernel, 0, (void *)h_inData);
-  checkError(status, "Failed to set fetch1 kernel arg");
-
-  status=clSetKernelArg(ffta_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set ffta kernel arg");
-  status=clSetKernelArg(fftb_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set fftb kernel arg");
-
-  // kernel stores to DDR memory
-  status=clSetKernelArg(store1_kernel, 0, sizeof(cl_mem), (void*)&d_inOutData);
-  checkError(status, "Failed to set store1 kernel arg");
-
-  // kernel fetches from DDR memory
-  status=clSetKernelArg(fetch2_kernel, 0, sizeof(cl_mem), (void*)&d_inOutData);
-  checkError(status, "Failed to set fetch2 kernel arg");
-  status=clSetKernelArg(fftc_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set fftc kernel arg");
-
-  // kernel stores using SVM based PCIe to host
-  status = clSetKernelArgSVMPointer(store2_kernel, 0, (void*)h_outData);
-  checkError(status, "Failed to set store2 kernel arg");
-
-  fft_time.exec_t = getTimeinMilliSec();
-  double first_half = getTimeinMilliSec();
-  status = clEnqueueTask(queue1, fetch1_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fetch kernel");
-
-  status = clEnqueueTask(queue2, ffta_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fft kernel");
-
-  status = clEnqueueTask(queue3, transpose_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch transpose kernel");
-
-  status = clEnqueueTask(queue4, fftb_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch second fft kernel");
-
-  status = clEnqueueTask(queue5, store1_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch second transpose kernel");
-
-  status = clFinish(queue1);
-  checkError(status, "failed to finish");
-  status = clFinish(queue2);
-  checkError(status, "failed to finish");
-  status = clFinish(queue3);
-  checkError(status, "failed to finish");
-  status = clFinish(queue4);
-  checkError(status, "failed to finish");
-  status = clFinish(queue5);
-  checkError(status, "failed to finish");
-  first_half = getTimeinMilliSec() - first_half;
-
-  double second_half = getTimeinMilliSec();
-  // enqueue fetch to same queue as the store kernel due to data dependency
-  status = clEnqueueTask(queue5, fetch2_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fetch kernel");
-
-  status = clEnqueueTask(queue4, fftc_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fft kernel");
-
-  status = clEnqueueTask(queue3, store2_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch transpose kernel");
-
-  status = clFinish(queue5);
-  checkError(status, "failed to finish");
-  status = clFinish(queue4);
-  checkError(status, "failed to finish");
-  status = clFinish(queue3);
-  checkError(status, "failed to finish");
-  /*
-  status = clFinish(queue2);
-  checkError(status, "failed to finish");
-  status = clFinish(queue1);
-  checkError(status, "failed to finish");
-  */
-
-  second_half = getTimeinMilliSec() - second_half;
-  fft_time.exec_t = getTimeinMilliSec() - fft_time.exec_t;
-  
-  printf("First half: %lf Second half: %lf\n\n", first_half, second_half);
-
-  status = clEnqueueSVMMap(queue1, CL_TRUE, CL_MAP_READ,
-    (void *)h_outData, sizeof(float2) * num_pts, 0, NULL, NULL);
-  checkError(status, "Failed to map out data");
-
-  for(size_t i = 0; i < num_pts; i++){
-    out[i].x = h_outData[i].x;
-    out[i].y = h_outData[i].y;
-  }
-
-  status = clEnqueueSVMUnmap(queue1, (void *)h_outData, 0, NULL, NULL);
-  checkError(status, "Failed to unmap out data");
-
-  if (h_inData)
-    clSVMFree(context, h_inData);
-  if (h_outData)
-    clSVMFree(context, h_outData);
-
-  queue_cleanup();
-
-  if (d_inOutData)
-    clReleaseMemObject(d_inOutData);
-
-  if(fetch1_kernel) 
-    clReleaseKernel(fetch1_kernel);  
-  if(fetch2_kernel) 
-    clReleaseKernel(fetch2_kernel);  
-
-  if(ffta_kernel) 
-    clReleaseKernel(ffta_kernel);  
-  if(fftb_kernel) 
-    clReleaseKernel(fftb_kernel);  
-  if(fftc_kernel) 
-    clReleaseKernel(fftc_kernel);  
-
-  if(transpose_kernel) 
-    clReleaseKernel(transpose_kernel);  
-
-  if(store1_kernel) 
-    clReleaseKernel(store1_kernel);  
-  if(store2_kernel) 
-    clReleaseKernel(store2_kernel);  
-
-  fft_time.valid = 1;
-  return fft_time;
-}
-
-/**
- * \brief  compute an out-of-place single precision complex 3D-FFT using the DDR of the FPGA for 3D Transpose
- * \param  N    : integer pointer addressing the size of FFT3d  
- * \param  inp  : float2 pointer to input data of size [N * N * N]
- * \param  out  : float2 pointer to output data of size [N * N * N]
- * \param  inv  : int toggle to activate backward FFT
- * \param  interleaving : 1 if using burst interleaved global memory buffers
- * \return fpga_t : time taken in milliseconds for data transfers and execution
- */
-fpga_t fftfpgaf_c2c_3d_ddr(int N, float2 *inp, float2 *out, bool inv) {
-  fpga_t fft_time = {0.0, 0.0, 0.0, 0};
-  cl_int status = 0;
-  int num_pts = N * N * N;
-  
-  // if N is not a power of 2
-  if(inp == NULL || out == NULL || ( (N & (N-1)) !=0)){
-    return fft_time;
-  }
-
-#ifdef VERBOSE
-  printf("Launching%s 3d FFT transform in DDR \n", inv ? " inverse":"");
-#endif
-
-  // Can't pass bool to device, so convert it to int
-  int inverse_int = (int)inv;
-
-  // Setup kernels
-  cl_kernel fetch1_kernel = clCreateKernel(program, "fetchBitrev1", &status);
-  checkError(status, "Failed to create fetch1 kernel");
-  cl_kernel ffta_kernel = clCreateKernel(program, "fft3da", &status);
-  checkError(status, "Failed to create fft3da kernel");
-  cl_kernel transpose_kernel = clCreateKernel(program, "transpose", &status);
-  checkError(status, "Failed to create transpose kernel");
-  cl_kernel fftb_kernel = clCreateKernel(program, "fft3db", &status);
-  checkError(status, "Failed to create fft3db kernel");
-  cl_kernel store1_kernel = clCreateKernel(program, "transposeStore1", &status);
-  checkError(status, "Failed to create store1 kernel");
-
-  cl_kernel fetch2_kernel = clCreateKernel(program, "fetchBitrev2", &status);
-  checkError(status, "Failed to create fetch2 kernel");
-  cl_kernel fftc_kernel = clCreateKernel(program, "fft3dc", &status);
-  checkError(status, "Failed to create fft3dc kernel");
-  cl_kernel store2_kernel = clCreateKernel(program, "transposeStore2", &status);
-  checkError(status, "Failed to create store2 kernel");
-
-  // Setup Queues to the kernels
-  queue_setup();
-
-  // Device memory buffers
-  cl_mem d_inData, d_transpose, d_outData;
-  d_inData = clCreateBuffer(context, CL_MEM_READ_ONLY | CL_CHANNEL_1_INTELFPGA, sizeof(float2) * num_pts, NULL, &status);
-  checkError(status, "Failed to allocate input device buffer\n");
-
-  d_transpose = clCreateBuffer(context, CL_MEM_READ_WRITE | CL_CHANNEL_2_INTELFPGA, sizeof(float2) * num_pts, NULL, &status);
-  checkError(status, "Failed to allocate output device buffer\n");
-
-  d_outData = clCreateBuffer(context, CL_MEM_WRITE_ONLY | CL_CHANNEL_1_INTELFPGA, sizeof(float2) * num_pts, NULL, &status);
-  checkError(status, "Failed to allocate output device buffer\n");
-
-  // Copy data from host to device
-  fft_time.pcie_write_t = getTimeinMilliSec();
-
-  status = clEnqueueWriteBuffer(queue1, d_inData, CL_TRUE, 0, sizeof(float2) * num_pts, inp, 0, NULL, NULL);
-
-  status = clFinish(queue1);
-  checkError(status, "failed to finish");
-
-  fft_time.pcie_write_t = getTimeinMilliSec() - fft_time.pcie_write_t;
-  checkError(status, "Failed to copy data to device");
-
-  status=clSetKernelArg(fetch1_kernel, 0, sizeof(cl_mem), (void *)&d_inData);
-  checkError(status, "Failed to set fetch1 kernel arg");
-
-  status=clSetKernelArg(ffta_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set ffta kernel arg");
-  status=clSetKernelArg(fftb_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set fftb kernel arg");
-
-  status=clSetKernelArg(store1_kernel, 0, sizeof(cl_mem), (void *)&d_transpose);
-  checkError(status, "Failed to set store1 kernel arg");
-
-  status=clSetKernelArg(fetch2_kernel, 0, sizeof(cl_mem), (void *)&d_transpose);
-  checkError(status, "Failed to set fetch2 kernel arg");
-  status=clSetKernelArg(fftc_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set fftc kernel arg");
-  status=clSetKernelArg(store2_kernel, 0, sizeof(cl_mem), (void *)&d_outData);
-  checkError(status, "Failed to set store2 kernel arg");
-
-  fft_time.exec_t = getTimeinMilliSec();
-  status = clEnqueueTask(queue1, fetch1_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fetch kernel");
-
-  status = clEnqueueTask(queue2, ffta_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fft kernel");
-
-  status = clEnqueueTask(queue3, transpose_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch transpose kernel");
-
-  status = clEnqueueTask(queue4, fftb_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch second fft kernel");
-
-  status = clEnqueueTask(queue5, store1_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch second transpose kernel");
-
-  // enqueue fetch to same queue as the store kernel due to data dependency
-  status = clEnqueueTask(queue5, fetch2_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fetch kernel");
-
-  status = clEnqueueTask(queue4, fftc_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fft kernel");
-
-  status = clEnqueueTask(queue3, store2_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch transpose kernel");
-
-  status = clFinish(queue5);
-  checkError(status, "failed to finish");
-  status = clFinish(queue4);
-  checkError(status, "failed to finish");
-  status = clFinish(queue3);
-  checkError(status, "failed to finish");
-  status = clFinish(queue2);
-  checkError(status, "failed to finish");
-  status = clFinish(queue1);
-  checkError(status, "failed to finish");
-
-  fft_time.exec_t = getTimeinMilliSec() - fft_time.exec_t;
-
-  // Copy results from device to host
-  fft_time.pcie_read_t = getTimeinMilliSec();
-  status = clEnqueueReadBuffer(queue1, d_outData, CL_TRUE, 0, sizeof(float2) * num_pts, out, 0, NULL, NULL);
-  
-  status = clFinish(queue1);
-  checkError(status, "failed to finish reading DDR using PCIe");
-
-  fft_time.pcie_read_t = getTimeinMilliSec() - fft_time.pcie_read_t;
-  checkError(status, "Failed to copy data from device");
-
-  queue_cleanup();
-
-  if (d_inData)
-    clReleaseMemObject(d_inData);
-  if (d_outData) 
-    clReleaseMemObject(d_outData);
-  if (d_transpose) 
-    clReleaseMemObject(d_transpose);
-
-  if(fetch1_kernel) 
-    clReleaseKernel(fetch1_kernel);  
-  if(fetch2_kernel) 
-    clReleaseKernel(fetch2_kernel);  
-
-  if(ffta_kernel) 
-    clReleaseKernel(ffta_kernel);  
-  if(fftb_kernel) 
-    clReleaseKernel(fftb_kernel);  
-  if(fftc_kernel) 
-    clReleaseKernel(fftc_kernel);  
-
-  if(transpose_kernel) 
-    clReleaseKernel(transpose_kernel);  
-
-  if(store1_kernel) 
-    clReleaseKernel(store1_kernel);  
-  if(store2_kernel) 
-    clReleaseKernel(store2_kernel);  
-
-  fft_time.valid = 1;
-  return fft_time;
+void queue_cleanup() {
+  if(queue1) 
+    clReleaseCommandQueue(queue1);
+  if(queue2) 
+    clReleaseCommandQueue(queue2);
+  if(queue3) 
+    clReleaseCommandQueue(queue3);
+  if(queue4) 
+    clReleaseCommandQueue(queue4);
+  if(queue5) 
+    clReleaseCommandQueue(queue5);
+  if(queue6) 
+    clReleaseCommandQueue(queue6);
+  if(queue7) 
+    clReleaseCommandQueue(queue7);
+  if(queue8) 
+    clReleaseCommandQueue(queue8);
 }
 
 /**
@@ -1544,6 +204,7 @@ fpga_t fftfpgaf_c2c_3d_ddr(int N, float2 *inp, float2 *out, bool inv) {
  * \param how_many : number of batched computations
  * \return fpga_t : time taken in milliseconds for data transfers and execution
  */
+/*
 fpga_t fftfpgaf_c2c_3d_ddr_batch(int N, float2 *inp, float2 *out, bool inv, bool interleaving, int how_many) {
   fpga_t fft_time = {0.0, 0.0, 0.0, 0};
   cl_int status = 0;
@@ -1677,7 +338,6 @@ fpga_t fftfpgaf_c2c_3d_ddr_batch(int N, float2 *inp, float2 *out, bool inv, bool
   clReleaseEvent(write_event[0]);
 
   // Loop over the 3 stages
-  
   for(size_t i = 2; i < how_many; i++){
 
     // Unblocking transfers between DDR and host 
@@ -1870,347 +530,4 @@ fpga_t fftfpgaf_c2c_3d_ddr_batch(int N, float2 *inp, float2 *out, bool inv, bool
   fft_time.valid = 1;
   return fft_time;
 }
-
-/**
- * \brief compute an batched out-of-place single precision complex 3D-FFT using the DDR of the FPGA for 3D Transpose and for data transfers between host's main memory and FPGA using Shared Virtual Memory 
- * \param N    : integer pointer addressing the size of FFT3d  
- * \param inp  : float2 pointer to input data of size [N * N * N]
- * \param out  : float2 pointer to output data of size [N * N * N]
- * \param inv  : int toggle to activate backward FFT
- * \param how_many : number of batched computations
- * \return fpga_t : time taken in milliseconds for data transfers and execution
- */
-fpga_t fftfpgaf_c2c_3d_ddr_svm_batch(int N, float2 *inp, float2 *out, bool inv, int how_many) {
-  fpga_t fft_time = {0.0, 0.0, 0.0, 0};
-  cl_int status = 0;
-  int num_pts = N * N * N;
-
-  // if N is not a power of 2
-  if(inp == NULL || out == NULL || ( (N & (N-1)) !=0) || (how_many <= 0)){
-    return fft_time;
-  }
-
-  if(!svm_enabled){
-    return fft_time;
-  }
-
-#ifdef VERBOSE
-  printf("Launching%s 3d FFT transform in DDR \n", inv ? " inverse":"");
-#endif
-
-  // Can't pass bool to device, so convert it to int
-  int inverse_int = (int)inv;
-
-  // Setup kernels
-  cl_kernel fetch1_kernel = clCreateKernel(program, "fetchBitrev1", &status);
-  checkError(status, "Failed to create fetch1 kernel");
-  cl_kernel ffta_kernel = clCreateKernel(program, "fft3da", &status);
-  checkError(status, "Failed to create fft3da kernel");
-  cl_kernel transpose_kernel = clCreateKernel(program, "transpose", &status);
-  checkError(status, "Failed to create transpose kernel");
-  cl_kernel fftb_kernel = clCreateKernel(program, "fft3db", &status);
-  checkError(status, "Failed to create fft3db kernel");
-  cl_kernel store1_kernel = clCreateKernel(program, "transposeStore1", &status);
-  checkError(status, "Failed to create store1 kernel");
-
-  cl_kernel fetch2_kernel = clCreateKernel(program, "fetchBitrev2", &status);
-  checkError(status, "Failed to create fetch2 kernel");
-  cl_kernel fftc_kernel = clCreateKernel(program, "fft3dc", &status);
-  checkError(status, "Failed to create fft3dc kernel");
-  cl_kernel store2_kernel = clCreateKernel(program, "transposeStore2", &status);
-  checkError(status, "Failed to create store2 kernel");
-
-  // Setup Queues to the kernels
-  queue_setup();
-
-  // Device memory buffers: double buffers
-  cl_mem d_outData_0;
-  d_outData_0 = clCreateBuffer(context, CL_MEM_READ_WRITE | CL_CHANNEL_1_INTELFPGA, sizeof(float2) * num_pts, NULL, &status);
-  checkError(status, "Failed to allocate output device buffer\n");
-
-  cl_mem d_outData_1;
-  d_outData_1 = clCreateBuffer(context, CL_MEM_READ_WRITE | CL_CHANNEL_2_INTELFPGA, sizeof(float2) * num_pts, NULL, &status);
-  checkError(status, "Failed to allocate output device buffer\n");
-
-  // allocate and initialize SVM buffers
-
-  float2 *h_inData[how_many], *h_outData[how_many];
-  for(size_t i = 0; i < how_many; i++){
-    h_inData[i] = (float2 *)clSVMAlloc(context, CL_MEM_READ_ONLY, sizeof(float2) * num_pts, 0);
-    h_outData[i] = (float2 *)clSVMAlloc(context, CL_MEM_WRITE_ONLY, sizeof(float2) * num_pts, 0);
-
-    status = clEnqueueSVMMap(queue1, CL_TRUE, CL_MAP_WRITE, (void *)h_inData[i], sizeof(float2) * num_pts, 0, NULL, NULL);
-    checkError(status, "Failed to map input data");
-
-    // copy data into h_inData
-    size_t stride = i * num_pts; 
-    for(size_t j = 0; j < num_pts; j++){
-      h_inData[i][j].x = inp[stride + j].x;
-      h_inData[i][j].y = inp[stride + j].y;
-    }
-
-    status = clEnqueueSVMUnmap(queue1, (void *)h_inData[i], 0, NULL, NULL);
-    checkError(status, "Failed to unmap input data");
-
-    status = clEnqueueSVMMap(queue1, CL_TRUE, CL_MAP_WRITE, (void *)h_outData[i], sizeof(float2) * num_pts, 0, NULL, NULL);
-    checkError(status, "Failed to map input data");
-
-    // copy data into h_inData
-    for(size_t j = 0; j < num_pts; j++){
-      h_outData[i][j].x = 0.0;
-      h_outData[i][j].y = 0.0;
-    }
-
-    status = clEnqueueSVMUnmap(queue1, (void *)h_outData[i], 0, NULL, NULL);
-    checkError(status, "Failed to unmap input data");
-  }
-
-  /*
-  * kernel arguments
-  */
-  // write to fetch kernel using SVM based PCIe
-  status = clSetKernelArgSVMPointer(fetch1_kernel, 0, (void *)h_inData[0]);
-  checkError(status, "Failed to set fetch1 kernel arg");
-
-  status=clSetKernelArg(ffta_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set ffta kernel arg");
-  status=clSetKernelArg(fftb_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set fftb kernel arg");
-
-  // kernel stores to DDR memory
-  status=clSetKernelArg(store1_kernel, 0, sizeof(cl_mem), (void *)&d_outData_0);
-  checkError(status, "Failed to set store1 kernel arg");
-
-  /*
-  *  First batch write phase
-  */
-  fft_time.exec_t = getTimeinMilliSec();
-  status = clEnqueueTask(queue1, fetch1_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fetch kernel");
-
-  status = clEnqueueTask(queue2, ffta_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fft kernel");
-
-  status = clEnqueueTask(queue3, transpose_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch transpose kernel");
-
-  status = clEnqueueTask(queue4, fftb_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch second fft kernel");
-
-  status = clEnqueueTask(queue5, store1_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch second transpose kernel");
-
-  for(size_t i = 1; i < how_many; i++){
-
-    /*
-    *  Read phase of previous iteration
-    */
-    // kernel fetches from DDR memory
-    // kernel stores using SVM based PCIe to host
-    if( (i % 2) == 1){
-      // if odd number of batches
-      status=clSetKernelArg(fetch2_kernel, 0, sizeof(cl_mem), (void *)&d_outData_0);
-      checkError(status, "Failed to set fetch2 kernel arg");
-
-      // Start fetch2 phase with same queue as store1
-      status = clEnqueueTask(queue5, fetch2_kernel, 0, NULL, NULL);
-      checkError(status, "Failed to launch fetch kernel");
-    }
-    else{
-      status=clSetKernelArg(fetch2_kernel, 0, sizeof(cl_mem), (void *)&d_outData_1);
-      checkError(status, "Failed to set fetch2 kernel arg");
-
-      // Start fetch2 phase with same queue as store1
-      status = clEnqueueTask(queue8, fetch2_kernel, 0, NULL, NULL);
-      checkError(status, "Failed to launch fetch kernel");
-    }
-    status=clSetKernelArg(fftc_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-    checkError(status, "Failed to set fftc kernel arg");
-    status = clSetKernelArgSVMPointer(store2_kernel, 0, (void *)h_outData[i-1]);
-    checkError(status, "Failed to set store2 kernel arg");
-
-    status = clEnqueueTask(queue6, fftc_kernel, 0, NULL, NULL);
-    checkError(status, "Failed to launch fft kernel");
-
-    status = clEnqueueTask(queue7, store2_kernel, 0, NULL, NULL);
-    checkError(status, "Failed to launch transpose kernel");
-
-    /*
-     *  write phase of current iteration
-     */
-    // change write phase host and ddr ptrs
-    status = clSetKernelArgSVMPointer(fetch1_kernel, 0, (void *)h_inData[i]);
-    checkError(status, "Failed to set fetch1 kernel arg");
-    if(i % 2 == 1){
-      status=clSetKernelArg(store1_kernel, 0, sizeof(cl_mem), (void *)&d_outData_1);
-      checkError(status, "Failed to set store1 kernel arg");
-    }
-    else{
-      status=clSetKernelArg(store1_kernel, 0, sizeof(cl_mem), (void *)&d_outData_0);
-      checkError(status, "Failed to set store1 kernel arg");
-    }
-
-    // Start write phase of current iteration
-    status = clEnqueueTask(queue1, fetch1_kernel, 0, NULL, NULL);
-    checkError(status, "Failed to launch fetch kernel");
-
-    status = clEnqueueTask(queue2, ffta_kernel, 0, NULL, NULL);
-    checkError(status, "Failed to launch fft kernel");
-
-    status = clEnqueueTask(queue3, transpose_kernel, 0, NULL, NULL);
-    checkError(status, "Failed to launch transpose kernel");
-
-    status = clEnqueueTask(queue4, fftb_kernel, 0, NULL, NULL);
-    checkError(status, "Failed to launch second fft kernel");
-
-    if(i % 2 == 1){
-      status = clEnqueueTask(queue8, store1_kernel, 0, NULL, NULL);
-      checkError(status, "Failed to launch second transpose kernel");
-    }
-    else{
-      status = clEnqueueTask(queue5, store1_kernel, 0, NULL, NULL);
-      checkError(status, "Failed to launch second transpose kernel");
-    }
-  }
-  
-  if(how_many % 2 == 1){
-    status=clSetKernelArg(fetch2_kernel, 0, sizeof(cl_mem), (void *)&d_outData_0);
-    checkError(status, "Failed to set fetch2 kernel arg");
-
-    status = clEnqueueTask(queue5, fetch2_kernel, 0, NULL, NULL);
-    checkError(status, "Failed to launch fetch kernel");
-  }
-  else{
-    status=clSetKernelArg(fetch2_kernel, 0, sizeof(cl_mem), (void *)&d_outData_1);
-    checkError(status, "Failed to set fetch2 kernel arg");
-    status = clEnqueueTask(queue8, fetch2_kernel, 0, NULL, NULL);
-    checkError(status, "Failed to launch fetch kernel");
-  }
-  status=clSetKernelArg(fftc_kernel, 0, sizeof(cl_int), (void*)&inverse_int);
-  checkError(status, "Failed to set fftc kernel arg");
-  status = clSetKernelArgSVMPointer(store2_kernel, 0, (void *)h_outData[how_many-1]);
-  checkError(status, "Failed to set store2 kernel arg");
-
-  status = clEnqueueTask(queue6, fftc_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch fft kernel");
-
-  status = clEnqueueTask(queue7, store2_kernel, 0, NULL, NULL);
-  checkError(status, "Failed to launch transpose kernel");
-
-  status = clFinish(queue1);
-  checkError(status, "Failed to finish queue1");
-  status = clFinish(queue2);
-  checkError(status, "Failed to finish queue2");
-  status = clFinish(queue3);
-  checkError(status, "Failed to finish queue1");
-  status = clFinish(queue4);
-  checkError(status, "Failed to finish queue2");
-  status = clFinish(queue5);
-  checkError(status, "Failed to finish queue1");
-  status = clFinish(queue6);
-  checkError(status, "Failed to finish queue2");
-  status = clFinish(queue7);
-  checkError(status, "Failed to finish queue1");
-  status = clFinish(queue8);
-  checkError(status, "Failed to finish queue2");
-
-  fft_time.exec_t = getTimeinMilliSec() - fft_time.exec_t;
-
-  for(size_t i = 0; i < how_many; i++){
-
-    status = clEnqueueSVMMap(queue2, CL_TRUE, CL_MAP_READ,
-      (void *)h_outData[i], sizeof(float2) * num_pts, 0, NULL, NULL);
-    checkError(status, "Failed to map out data");
-
-    size_t stride = i * num_pts;
-    for(size_t j = 0; j < num_pts; j++){
-      out[stride + j].x = h_outData[i][j].x;
-      out[stride + j].y = h_outData[i][j].y;
-    }
-
-    status = clEnqueueSVMUnmap(queue2, (void *)h_outData[i], 0, NULL, NULL);
-    checkError(status, "Failed to unmap out data");
-  }
-
-  for(size_t i = 0; i < how_many; i++){
-    clSVMFree(context, h_inData[i]);
-    clSVMFree(context, h_outData[i]);
-  }
-
-  queue_cleanup();
-
-  if (d_outData_0) 
-    clReleaseMemObject(d_outData_0);
-  if (d_outData_1) 
-    clReleaseMemObject(d_outData_1);
-
-  if(fetch1_kernel) 
-    clReleaseKernel(fetch1_kernel);  
-  if(fetch2_kernel) 
-    clReleaseKernel(fetch2_kernel);  
-
-  if(ffta_kernel) 
-    clReleaseKernel(ffta_kernel);  
-  if(fftb_kernel) 
-    clReleaseKernel(fftb_kernel);  
-  if(fftc_kernel) 
-    clReleaseKernel(fftc_kernel);  
-
-  if(transpose_kernel) 
-    clReleaseKernel(transpose_kernel);  
-
-  if(store1_kernel) 
-    clReleaseKernel(store1_kernel);  
-  if(store2_kernel) 
-    clReleaseKernel(store2_kernel);  
-
-  fft_time.valid = 1;
-  return fft_time;
-}
-
-
-/**
- * \brief Create a command queue for each kernel
- */
-void queue_setup(){
-  cl_int status = 0;
-  // Create one command queue for each kernel.
-  queue1 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
-  checkError(status, "Failed to create command queue1");
-  queue2 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
-  checkError(status, "Failed to create command queue2");
-  queue3 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
-  checkError(status, "Failed to create command queue3");
-  queue4 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
-  checkError(status, "Failed to create command queue4");
-  queue5 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
-  checkError(status, "Failed to create command queue5");
-  queue6 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
-  checkError(status, "Failed to create command queue6");
-  queue7 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
-  checkError(status, "Failed to create command queue6");
-  queue8 = clCreateCommandQueue(context, device, CL_QUEUE_PROFILING_ENABLE, &status);
-  checkError(status, "Failed to create command queue6");
-}
-
-/**
- * \brief Release all command queues
- */
-void queue_cleanup() {
-  if(queue1) 
-    clReleaseCommandQueue(queue1);
-  if(queue2) 
-    clReleaseCommandQueue(queue2);
-  if(queue3) 
-    clReleaseCommandQueue(queue3);
-  if(queue4) 
-    clReleaseCommandQueue(queue4);
-  if(queue5) 
-    clReleaseCommandQueue(queue5);
-  if(queue6) 
-    clReleaseCommandQueue(queue6);
-  if(queue7) 
-    clReleaseCommandQueue(queue7);
-  if(queue8) 
-    clReleaseCommandQueue(queue8);
-}
+*/
